+++ libs/ui/src/components/atoms/Link/Link.stories.tsx
@@ -40,26 +40,3 @@ InternalLinkWithIcon.args = {
   children: 'Internal Link',
   Icon: RiArrowDropDownLine,
 };
-
-export const NavigationLink = Template.bind({});
-NavigationLink.args = {
-  href: '/Home',
-  linkType: 'navigation',
-  children: 'Navigation Link',
-};
-
-export const NavigationLinkWithIcon = Template.bind({});
-NavigationLinkWithIcon.args = {
-  href: '/Home',
-  linkType: 'navigation',
-  children: 'Navigation Link with Icon',
-  Icon: RiArrowDropDownLine,
-};
-
-export const NavigationLinkSelected = Template.bind({});
-NavigationLinkSelected.args = {
-  href: '/Home',
-  linkType: 'navigation',
-  children: 'Navigation Link Selected',
-  selected: true,
-};
