+++ libs/ui/src/components/atoms/Link/Link.styles.ts
@@ -30,32 +30,3 @@ export const ExternalLink = styled.a`
 export const InternalLink = styled(Link)`
   ${LinkStyles}
 `;
-
-export const NavLink = styled(Link)`
-  ${LinkStyles};
-  border-bottom: 2px transparent solid;
-  color: ${(props) => props.theme.subNav.navLinkColor};
-  cursor: pointer;
-  font-size: 2.4rem;
-  letter-spacing: 1.5px;
-  padding-bottom: 1rem;
-  transition: 0.2s all;
-
-  svg {
-    margin-left: 0.3rem;
-  }
-
-  :hover {
-    color: ${(props) => props.theme.subNav.navLinkHoverColor};
-    text-decoration: none;
-  }
-
-  &.selected {
-    color: white;
-    border-bottom: 2px ${(props) => props.theme.subNav.navLinkColor} solid;
-  }
-
-  &.subnav {
-    padding-bottom: 1rem;
-  }
-`;
